sensor: adxl345: reorganize defined fields and macros
Rename defines to make it more consistend and remove duplicates. Make it
clear if we use an ADXL345_REG_ for an entire register, or just a single
bit. Apply bitfield macros, such as BIT(), GENMASK(), FIELD_GET(), etc.
to factor out plain logic operations. Help out readability and
maintainability. This step is a preparation for upcoming implementation on
this sensor driver.

Signed-off-by: Lothar Rubusch <l.rubusch@gmail.com>