sensor: adxl345: update fetch and get samples
Update the implementation of fetch and get. Default to FIFO BYPASS mode.
When GPIO lines are defined in the DT and CONFIG_ADXL345_TRIGGER is set,
Analog's FIFO BYPASS mode is configured. In this case up to watermark
samples are stored in the FIFO then a watermark event is triggered. If
samples are not read out, an overrun event can be configured. Aside, also
data ready event can be sent out by the sensor.

Signed-off-by: Lothar Rubusch <l.rubusch@gmail.com>