sensor: adxl345: refactor ODR initialization
Currently, the ODR supposed to be configured in the devicetree, then
at the start of the adxl345_init() re-written to hardcoded
ADXL345_RATE_25HZ, and at the end of the adxl345_init() set again to
data->odr. Why?!

Refactor sensor ODR initialization and define a reasonable default in case
devicetree configuration fails. In adxl345_init(), use the ODR value
already provided in the config to update the corresponding register,
ensuring that only the relevant field is modified.

Use FIELD_GET() on the ODR mask when setting the value, and write directly
to the register both during sensor initialization and within the attribute
function. This removes an unnecessary layer of indirection, allowing the
setter function to be dropped entirely.

Signed-off-by: Lothar Rubusch <l.rubusch@gmail.com>