sensor: adxl345: simplify event handler registration
For the simplified fetch-and-get implementation, clean up the trigger
source by removing the unnecessary re-mapping of interrupt lines left over
from the old trigger logic. Allow applications to register handlers for
data-ready, overrun, and watermark sensor events, as also activity.

Provide a mechanism for applications to register these handlers, and
alternatively offer default handling that clears and resets the FIFO after
an interrupt is received. Additionally, introduce a dedicated function to
flush and reset the FIFO.

Signed-off-by: Lothar Rubusch <l.rubusch@gmail.com>