sensor: adxl345: rename ADXL345_X_AXIS_DATA_0_REG to ADXL345_REG_DATA_XYZ_REGS
Rename this register for better clarity. According to the datasheet, this
register corresponds to the first x-axis data byte, followed by the second
x-axis byte, then the y-axis, and finally the z-axis data in a similar
sequence. The sensor does not provide an option to selectively enable or
disable specific axes for FIFO usage. Consequently, all three axis data
must be fetched together for a single FIFO sample line.

By using this register as the starting point for the fetch, it is always
used to retrieve x-, y-, and z-axis samples in a single operation.
Therefore, renaming it to ADXL345_REG_DATA_XYZ_REGS would be more
appropriate and meaningful for the driver implementation.

Signed-off-by: Lothar Rubusch <l.rubusch@gmail.com>