sensor: adxl345: DTS: set minimum watermark to 2
The FIFO is used either to buffer values in Analog's "stream mode" or to
record a set of measurements triggered by a specific sensor event in
Analog's "trigger mode". Since sensor event handling is not yet implemented
in this driver, only stream mode is applicable. In both cases - especially
in stream mode - setting the watermark to 1 is not practical. Doing so
would generate an interrupt for every single measured value. If capturing
every measurement in real time is the goal, Analog's "bypass mode" is more
appropriate, as it avoids the additional overhead of handling interrupts
like watermark or data ready. Moreover, generating frequent interrupts can
easily cause overrun issues.

Therefore, the minimum configurable watermark should be changed to 2.

Signed-off-by: Lothar Rubusch <l.rubusch@gmail.com>